Add db-config-ini-key to CLI options
This allows us to use the database connection string in a config file
that does not use "db-connection-string" as the key.  For example in a
config file like this:

```
[app:main]
postgresql.db-connection-string = dbname=travis user=travis host=localhost
```

dbmigrator will use this setting if
`--db-config-ini-key postgresql.db-connection-string` is passed in.

Close #5