Add `--sort` (version or applied) to `dbmigrator list`
`dbmigrator list` only shows migrations in the order of migration
filenames at the moment.  It works well for when users are trying to
migrate.  When users are trying to rollback however, it is difficult to
tell which migration will be rolled back because the applied dates are
not in the same order as the migration filenames.

This change allows the user to specify the order by doing
`--sort=applied`:

```
version        | name            | is applied | date applied
----------------------------------------------------------------------
20160228212456   cool_stuff        True         2017-10-20 10:27:57.011677-07:00
20160228202637   add_table         False
```